+++ explore/src/main/scala/explore/config/ConfigurationTile.scala
@@ -3,6 +3,7 @@
 
 package explore.config
 
+import cats.effect.IO
 import cats.syntax.all.*
 import clue.FetchClient
 import clue.data.Assign
@@ -10,48 +11,46 @@ import clue.data.Input
 import clue.data.syntax.*
 import crystal.*
 import crystal.react.*
+import crystal.react.View
+import eu.timepit.refined.types.string.NonEmptyString
 import explore.*
 import explore.DefaultErrorPolicy
 import explore.common.Aligner
 import explore.common.ScienceQueries.ScienceRequirementsUndoView
 import explore.common.ScienceQueries.UpdateScienceRequirements
-import explore.model.AppContext
-import explore.model.ScienceRequirements.Spectroscopy
-import explore.model.itc.ItcTarget
-import japgolly.scalajs.react.*
-import japgolly.scalajs.react.vdom.html_<^.*
-import lucuma.react.common.ReactFnProps
-import lucuma.schemas.ObservationDB
-import lucuma.schemas.ObservationDB.Types.*
-import lucuma.schemas.model.BasicConfiguration
-import lucuma.schemas.odb.input.*
-import monocle.Iso
-import queries.common.ObsQueriesGQL
-
-import cats.effect.IO
-import crystal.react.View
-import eu.timepit.refined.types.string.NonEmptyString
 import explore.components.Tile
 import explore.components.ui.ExploreStyles
+import explore.model.AppContext
 import explore.model.AsterismIds
 import explore.model.InstrumentConfigAndItcResult
 import explore.model.ObsConfiguration
 import explore.model.ObsTabTileIds
 import explore.model.Observation
 import explore.model.ScienceRequirements
+import explore.model.ScienceRequirements.Spectroscopy
 import explore.model.TargetList
 import explore.model.enums.WavelengthUnits
+import explore.model.itc.ItcTarget
 import explore.modes.InstrumentConfig
 import explore.modes.SpectroscopyModesMatrix
 import explore.undo.*
+import japgolly.scalajs.react.*
 import japgolly.scalajs.react.Callback
+import japgolly.scalajs.react.vdom.html_<^.*
 import lucuma.core.model.CoordinatesAtVizTime
 import lucuma.core.model.PosAngleConstraint
 import lucuma.core.model.Program
 import lucuma.core.model.User
+import lucuma.react.common.ReactFnProps
+import lucuma.schemas.ObservationDB
+import lucuma.schemas.ObservationDB.Types.*
+import lucuma.schemas.model.BasicConfiguration
 import lucuma.schemas.model.ObservingMode
+import lucuma.schemas.odb.input.*
 import lucuma.ui.syntax.all.given
+import monocle.Iso
 import org.typelevel.log4cats.Logger
+import queries.common.ObsQueriesGQL
 import queries.schemas.itc.syntax.*
 
 object ConfigurationTile:
