+++ explore/src/main/scala/explore/itc/ItcTile.scala
@@ -8,29 +8,29 @@ import cats.syntax.all.*
 import crystal.*
 import crystal.react.*
 import crystal.react.hooks.*
+import eu.timepit.refined.*
 import eu.timepit.refined.types.string.NonEmptyString
+import explore.Icons
 import explore.common.UserPreferencesQueries
 import explore.common.UserPreferencesQueries.*
-import explore.model.AppContext
-import explore.model.itc.*
-import lucuma.core.util.NewType
-import lucuma.itc.GraphType
 import explore.components.Tile
+import explore.components.ui.ExploreStyles
+import explore.model.AppContext
+import explore.model.Constants.MissingInfoMsg
 import explore.model.GlobalPreferences
 import explore.model.ObsTabTileIds
 import explore.model.Observation
 import explore.model.TargetList
-import lucuma.core.model.User
-import eu.timepit.refined.*
-import explore.Icons
-import explore.components.ui.ExploreStyles
-import explore.model.Constants.MissingInfoMsg
+import explore.model.itc.*
 import explore.model.itc.ItcAsterismGraphResults
 import explore.model.itc.ItcGraphResult
 import explore.model.itc.ItcTarget
 import explore.model.reusability.given
 import japgolly.scalajs.react.*
 import japgolly.scalajs.react.vdom.html_<^.*
+import lucuma.core.model.User
+import lucuma.core.util.NewType
+import lucuma.itc.GraphType
 import lucuma.react.common.ReactFnProps
 import lucuma.react.floatingui.syntax.*
 import lucuma.react.primereact.Dropdown
