+++ explore/src/main/scala/explore/tabs/ObsTabContents.scala
@@ -29,11 +29,12 @@ import explore.model.enums.SelectedPanel
 import explore.model.reusability.given
 import explore.modes.SpectroscopyModesMatrix
 import explore.observationtree.*
+import explore.plots.ElevationPlotTile
+import explore.plots.ObjectPlotData
+import explore.plots.PlotData
 import explore.shortcuts.*
 import explore.shortcuts.given
 import explore.syntax.ui.*
-import explore.plots.ObjectPlotData
-import explore.plots.PlotData
 import explore.undo.UndoContext
 import explore.undo.UndoSetter
 import explore.utils.*
@@ -56,7 +57,6 @@ import lucuma.ui.reusability.given
 import lucuma.ui.sso.UserVault
 import lucuma.ui.syntax.all.given
 import monocle.Iso
-import explore.plots.ElevationPlotTile
 
 object DeckShown extends NewType[Boolean]:
   inline def Shown: DeckShown  = DeckShown(true)
