+++ explore/src/main/scala/explore/tabs/TargetTabContents.scala
@@ -31,15 +31,14 @@ import explore.model.enums.SelectedPanel
 import explore.model.reusability.given
 import explore.model.syntax.all.*
 import explore.observationtree.AsterismGroupObsList
+import explore.plots.ElevationPlotTile
+import explore.plots.ObjectPlotData
+import explore.plots.PlotData
 import explore.shortcuts.*
 import explore.shortcuts.given
 import explore.targeteditor.AsterismEditorTile
-import explore.plots.ObjectPlotData
-import explore.plots.PlotData
 import explore.targets.TargetPasteAction
-import explore.targets.TargetSummaryBody
-import explore.targets.TargetSummaryTileState
-import explore.targets.TargetSummaryTitle
+import explore.targets.TargetSummaryTile
 import explore.undo.*
 import explore.utils.*
 import japgolly.scalajs.react.*
@@ -64,7 +63,6 @@ import lucuma.ui.syntax.all.given
 import monocle.Iso
 import org.typelevel.log4cats.Logger
 import queries.schemas.odb.ObsQueries
-import explore.plots.ElevationPlotTile
 
 import java.time.Instant
 import scala.collection.immutable.SortedSet
@@ -334,13 +332,8 @@ object TargetTabContents extends TwoPanels:
           /**
            * Render the summary table.
            */
-          val renderSummary: Tile[TargetSummaryTileState] = Tile(
-            ObsTabTileIds.TargetSummaryId.id,
-            s"Target Summary (${props.targets.get.size})",
-            TargetSummaryTileState(Nil, none),
-            backButton.some
-          )(
-            TargetSummaryBody(
+          val renderSummary: Tile[TargetSummaryTile.TileState] =
+            TargetSummaryTile(
               props.userId,
               props.programId,
               props.targets.model,
@@ -350,10 +343,9 @@ object TargetTabContents extends TwoPanels:
               selectedTargetIds,
               props.focusedSummaryTargetId,
               focusTargetId,
-              _
-            ),
-            (s, _) => TargetSummaryTitle(props.programId, props.readonly, s)
-          )
+              props.readonly,
+              backButton
+            )
 
           val plotData: PlotData =
             PlotData:
